+++ js/assetReplacementManager.js
@@ -1,6 +1,6 @@
 import { GLTFLoader } from 'three/addons/loaders/GLTFLoader.js';
 import { Downloader } from './downloader.js';
-import { setupAnimatedPlayer, setupAnimatedRobot, setupAnimatedChicken, setupAnimatedWireframe, setupAnimatedAlien, setupEyebot, setupAnimatedShopkeeper, setupAnimatedOgre, setupAnimatedKnight } from './animationSetup.js';
+import { setupAnimatedPlayer, setupAnimatedRobot, setupAnimatedChicken, setupAnimatedWireframe, setupAnimatedAlien, setupEyebot, setupAnimatedShopkeeper, setupAnimatedOgre, setupAnimatedKnight, setupAnimatedSprite } from './animationSetup.js';
 
 export class AssetReplacementManager {
     /* @tweakable The number of models to parse in parallel. Higher values may be faster but use more memory and CPU. */
@@ -57,6 +57,12 @@ export class AssetReplacementManager {
                 setupFn: setupAnimatedShopkeeper,
                 applyFn: (modelData) => this.dependencies.npcManager.useAnimatedShopkeepers(modelData)
             },
+            "sprite": {
+                assetNames: ["Sprite Idle Animation", "Sprite Walking Animation", "Sprite Running Animation", "Sprite Listening Animation"],
+                clipNames: ["idle", "walk", "run", "listen"],
+                setupFn: setupAnimatedSprite,
+                applyFn: (modelData) => this.dependencies.npcManager.useAnimatedSprites(modelData)
+            },
             'ogre': {
                 assetNames: ['Ogre Idle Animation', 'Ogre Walking Animation', 'Ogre Running Animation', 'Ogre Listening Animation'],
                 clipNames: ['idle', 'walk', 'run', 'listen'],
@@ -64,8 +70,15 @@ export class AssetReplacementManager {
                 applyFn: (modelData) => this.dependencies.npcManager.useAnimatedOgres(modelData)
             },
             'knight': {
-                assetNames: ['Knight Idle Animation', 'Knight Walking Animation', 'Knight Running Animation', 'Knight Listening Animation'],
-                clipNames: ['idle', 'walk', 'run', 'listen'],
+                assetNames: [
+                    'Knight Idle Animation',
+                    'Knight Walking Animation',
+                    'Knight Running Animation',
+                    'Knight Listening Animation',
+                    'Knight Cheer with Both Hands Animation',
+                    'Knight Cheer with Both Hands 1 Animation'
+                ],
+                clipNames: ['idle', 'walk', 'run', 'listen', 'cheer', 'cheer1'],
                 setupFn: setupAnimatedKnight,
                 applyFn: (modelData) => this.dependencies.npcManager.useAnimatedKnights(modelData)
             }